pyproject.toml
[project]
name = "podcast-summary-tool"
version = "1.0.0"
description = "AI-powered podcast summary tool for extracting insights and actionable alpha"
authors = [{name = "Claude Code", email = "noreply@anthropic.com"}]
readme = "README.md"
requires-python = ">=3.8.1"
keywords = ["podcast", "ai", "analysis", "youtube", "transcription", "claude", "insights", "alpha"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "youtube-transcript-api>=0.6.2",
    "yt-dlp>=2024.8.6",
    "pdfkit>=1.0.0",
    "PyYAML>=6.0.1",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
    "html5lib>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
podcast-summary = "src.podcast_summary:main"
podcast-pdf = "src.converters.md_to_pdf:main"

[project.urls]
Homepage = "https://github.com/anthropics/claude-code"
Documentation = "https://docs.anthropic.com/en/docs/claude-code"
Repository = "https://github.com/anthropics/claude-code"
Issues = "https://github.com/anthropics/claude-code/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
asyncio_mode = "auto"